kea: move runtime dir from /run/kea to /run/kea{4,6}-${iface}
Kea may clear the runtime directory when starting (I don't know when it
started, maybe it was always like this when the interval between
service starts is too big). This erases some runtime files if you run
multiple instances with the same runtime dir at the same time. To
prevent this, move each Kea's runtime files into a separate directory.